MissionManagerFragment 완성 MissionCreateActivity 완성
#todo
Image첨부 미션 수행
Image첨부 미션 공유하기
미션 수행(Applicant) UI 및 Logic
마니또 종료시 내부 DB 삭제

+++ app/src/main/java/com/example/mnt_android/service/model/CheckRoom.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 
 class CheckRoom(
     @SerializedName("id")
-    val id: Int,
+    val id: Long,
     @SerializedName("isCreater")
     val isCreater: Int,
     @SerializedName("manittoId")
@@ -18,7 +18,7 @@ class CheckRoom(
     val user : User
 ) : Parcelable {
     constructor(parcel: Parcel) : this(
-        parcel.readInt(),
+        parcel.readLong(),
         parcel.readInt(),
         parcel.readString(),
         parcel.readParcelable(Room::class.java.classLoader),
@@ -27,7 +27,7 @@ class CheckRoom(
     }
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
-        parcel.writeInt(id)
+        parcel.writeLong(id)
         parcel.writeInt(isCreater)
         parcel.writeString(manittoId)
         parcel.writeParcelable(room, flags)
